requirements.txt
appdirs==1.4.4
attrs==20.1.0
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
idna==2.10
isodate==0.6.0
lxml==4.5.2
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.25.10
zeep==3.4.0
